Add files via upload
BeirRun is a fun and creative game, at its bear bones it is a two-dimensional birds eye view video game. The objective of the game is to collect drinks. However, as the player picks up drinks the controls get slightly harder, the graphics distort and the music/sounds will also distort. The main purpose of this game is to learn how to build an active community of developers and players. We hope to accomplish this by having open communications both ways, involving the community in adding features and removing bugs.